修改timeout为float，增加协程Redis timeout配置

+++ src/redis/src/AbstractRedisConnection.php
@@ -153,7 +153,7 @@ abstract class AbstractRedisConnection extends AbstractConnection
      *
      * @return Redis | \Redis
      */
-    abstract protected function getConnectRedis(string $host, int $port, int $timeout);
+    abstract protected function getConnectRedis(string $host, int $port, float $timeout);
 
     /**
      * @param string $method
